Fix infinite loop in neovim 0.5.2
There was an infinite loop with the 0.5.2+ vim.diagnostic API where
setting the coverage diagnostics would trigger the autocmd that would
re-trigger fetching & setting coverage. The new logic is to only update
coverage when the flow LSP client receives a publishDiagnostics event.
However, since the flow server will only publish diagnostics when an
error is introduced or removed, there is a new configurable interval to
fetch coverage. This allows the coverage numbers to update when making
changes that don't introduce or remove flow errors.

This commit also makes it so that no configuration is required (user
does not have to call on_attach)